// src/app/model/productionGeneralInfoRequest.ts
export interface ProductionGeneralInfoRequest { 
    initialProdDate?: Date;
    initialType?: ProductionGeneralInfoRequest.InitialTypeEnum;
    initialProduct?: ProductionGeneralInfoRequest.InitialProductEnum;
    initialLiftType?: ProductionGeneralInfoRequest.InitialLiftTypeEnum;
    initStatus?: ProductionGeneralInfoRequest.InitStatusEnum;
    monitoringSystem?: ProductionGeneralInfoRequest.MonitoringSystemEnum;
    currentWellType?: ProductionGeneralInfoRequest.CurrentWellTypeEnum;
    currentWellTypeDate?: Date;
    currentProduct?: ProductionGeneralInfoRequest.CurrentProductEnum;
    currentStatus?: ProductionGeneralInfoRequest.CurrentStatusEnum;
    runtime?: number;
    currentLiftType?: ProductionGeneralInfoRequest.CurrentLiftTypeEnum;
    currentLiftTypeDate?: Date;
    powerSourceType?: ProductionGeneralInfoRequest.PowerSourceTypeEnum;
    powerSource?: string;
    processionPlant?: ProductionGeneralInfoRequest.ProcessionPlantEnum;
}
export namespace ProductionGeneralInfoRequest {
    export type InitialTypeEnum = 'oil producer' | 'gas producer' | 'water producer' | 'oil  and gas' | 'condensate producer' | 'non';
    export const InitialTypeEnum = {
        OilProducer: 'oil producer' as InitialTypeEnum,
        GasProducer: 'gas producer' as InitialTypeEnum,
        WaterProducer: 'water producer' as InitialTypeEnum,
        OilAndGas: 'oil  and gas' as InitialTypeEnum,
        CondensateProducer: 'condensate producer' as InitialTypeEnum,
        Non: 'non' as InitialTypeEnum
    };
    export type InitialProductEnum = 'oil' | 'water' | 'gas' | 'condensate' | 'non condensate';
    export const InitialProductEnum = {
        Oil: 'oil' as InitialProductEnum,
        Water: 'water' as InitialProductEnum,
        Gas: 'gas' as InitialProductEnum,
        Condensate: 'condensate' as InitialProductEnum,
        NonCondensate: 'non condensate' as InitialProductEnum
    };
    export type InitialLiftTypeEnum = 'NaturalFlow' | 'S/R' | 'ESP' | 'Gl' | 'PCP' | 'PlungerLift' | 'ESP-PCP';
    export const InitialLiftTypeEnum = {
        NaturalFlow: 'NaturalFlow' as InitialLiftTypeEnum,
        SR: 'S/R' as InitialLiftTypeEnum,
        ESP: 'ESP' as InitialLiftTypeEnum,
        Gl: 'Gl' as InitialLiftTypeEnum,
        PCP: 'PCP' as InitialLiftTypeEnum,
        PlungerLift: 'PlungerLift' as InitialLiftTypeEnum,
        ESPPCP: 'ESP-PCP' as InitialLiftTypeEnum
    };
    export type InitStatusEnum = 'producing' | 'dry';
    export const InitStatusEnum = {
        Producing: 'producing' as InitStatusEnum,
        Dry: 'dry' as InitStatusEnum
    };
    export type MonitoringSystemEnum = 'skada' | 'non';
    export const MonitoringSystemEnum = {
        Skada: 'skada' as MonitoringSystemEnum,
        Non: 'non' as MonitoringSystemEnum
    };
    export type CurrentWellTypeEnum = 'oil producer' | 'gas producer' | 'water producer' | 'oil  and gas' | 'condensate producer' | 'non';
    export const CurrentWellTypeEnum = {
        OilProducer: 'oil producer' as CurrentWellTypeEnum,
        GasProducer: 'gas producer' as CurrentWellTypeEnum,
        WaterProducer: 'water producer' as CurrentWellTypeEnum,
        OilAndGas: 'oil  and gas' as CurrentWellTypeEnum,
        CondensateProducer: 'condensate producer' as CurrentWellTypeEnum,
        Non: 'non' as CurrentWellTypeEnum
    };
    export type CurrentProductEnum = 'oil' | 'water' | 'gas' | 'condensate' | 'non condensate';
    export const CurrentProductEnum = {
        Oil: 'oil' as CurrentProductEnum,
        Water: 'water' as CurrentProductEnum,
        Gas: 'gas' as CurrentProductEnum,
        Condensate: 'condensate' as CurrentProductEnum,
        NonCondensate: 'non condensate' as CurrentProductEnum
    };
    export type CurrentStatusEnum = 'producing' | 'shutin';
    export const CurrentStatusEnum = {
        Producing: 'producing' as CurrentStatusEnum,
        Shutin: 'shutin' as CurrentStatusEnum
    };
    export type CurrentLiftTypeEnum = 'NaturalFlow' | 'S/R' | 'ESP' | 'Gl' | 'PCP' | 'PlungerLift' | 'ESP-PCP';
    export const CurrentLiftTypeEnum = {
        NaturalFlow: 'NaturalFlow' as CurrentLiftTypeEnum,
        SR: 'S/R' as CurrentLiftTypeEnum,
        ESP: 'ESP' as CurrentLiftTypeEnum,
        Gl: 'Gl' as CurrentLiftTypeEnum,
        PCP: 'PCP' as CurrentLiftTypeEnum,
        PlungerLift: 'PlungerLift' as CurrentLiftTypeEnum,
        ESPPCP: 'ESP-PCP' as CurrentLiftTypeEnum
    };
    export type PowerSourceTypeEnum = 'OHL' | 'Diesel Generator';
    export const PowerSourceTypeEnum = {
        OHL: 'OHL' as PowerSourceTypeEnum,
        DieselGenerator: 'Diesel Generator' as PowerSourceTypeEnum
    };
    export type ProcessionPlantEnum = 'Gas Plant' | 'Oil Plant' | 'Flair';
    export const ProcessionPlantEnum = {
        GasPlant: 'Gas Plant' as ProcessionPlantEnum,
        OilPlant: 'Oil Plant' as ProcessionPlantEnum,
        Flair: 'Flair' as ProcessionPlantEnum
    };
}